requirements.txt
requests~=2.32.3
mysql-connector-python~=9.2.0
gspread~=6.2.0
tqdm~=4.67.1
oauth2client~=4.1.3